refactor: move gameHooksMap from lib/constants/ to hooks/shared/game-state/ to fix circular dep (#389)
lib/ must never import from app/. gameHooksMap imported useMathGame and
useCountingGame from app/games/ — same pattern fixed for tetris/building
stores in #377. Moving the file to hooks/ (which is allowed to import
from app/) resolves the violation. Update 2 import sites.

Closes #389
