+++ gamesformykids/hooks/shared/game-state/useGameConfig.ts
@@ -8,7 +8,7 @@
 import { useMemo } from 'react'
 import { GameType, BaseGameItem } from '@/lib/types/core/base'
 import { GAME_UI_CONFIGS } from '@/lib/constants/ui/gameConfigs'
-import { GAME_HOOKS_MAP, AutoGameType } from '@/lib/constants/gameHooksMap'
+import { GAME_HOOKS_MAP, AutoGameType } from './gameHooksMap'
 import { GameCardMap } from '@/components/shared/GameCardMap'
 import {
   GameConfigContextValue,
